Bugfix: Doors are no longer (sometimes) generated without a tile.
Turns out, rooms can share doors if they’re next to each other.  This
was causing duplicate doors to be spawned on top of each other.  Since
the first door to be generated in a cell would cause it to be impassable
the second door couldn’t be added to the tile.  Problems ensued.  Now
it’s fixed. 

Signed-off-by: suicidepills <atom.mac@gmail.com>